refactor: nested class/interfaceを別ファイルに分割
