+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/CustomAdapter.kt
@@ -1,16 +1,13 @@
 package jp.co.yumemi.android.codecheck
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 
 class CustomAdapter(
     private val itemClickListener: OnItemClickListener,
-) : ListAdapter<RepositoryInfo, CustomAdapter.ViewHolder>(object :
+) : ListAdapter<RepositoryInfo, ViewHolder>(object :
     DiffUtil.ItemCallback<RepositoryInfo>() {
     override fun areItemsTheSame(
         oldRepositoryInfo: RepositoryInfo,
@@ -26,18 +23,6 @@ class CustomAdapter(
         return oldRepositoryInfo == newRepositoryInfo
     }
 }) {
-
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        private val nameText = view.findViewById<TextView>(R.id.repositoryNameView)
-
-        fun bind(repositoryInfo: RepositoryInfo) {
-            nameText.text = repositoryInfo.name
-        }
-    }
-
-    interface OnItemClickListener {
-        fun itemClick(item: RepositoryInfo)
-    }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context)
