+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/OnItemClickListener.kt
@@ -1,0 +1,5 @@
+package jp.co.yumemi.android.codecheck
+
+interface OnItemClickListener {
+    fun itemClick(item: RepositoryInfo)
+}
